Combine Interface: IRQs; libcpu: inspired more bitflags
